Collapse TextureStitchEvent into a single post-stitch event
