+++ patches/net/minecraft/client/renderer/texture/TextureAtlas.java.patch
@@ -5,7 +5,7 @@
           this.animatedTextures = List.copyOf(list1);
        }
 +
-+      net.neoforged.neoforge.client.ClientHooks.onTextureStitchedPost(this);
++      net.neoforged.neoforge.client.ClientHooks.onTextureAtlasStitched(this);
     }
  
     @Override
